fix(conda environment): add C++ and R linting dependencies in conda
`pre-commit install` only handles some of the dependencies required for some hooks. others need to
be managed separately. this adds what I think are all the R requirements, and some of the C++
requirements. this doesn't add clang-format, which isn't present in the standard conda channels so
far as I can tell; not yet sure what to do about this.

re #1 #2